# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Drogaria",
        "description": "Esta documentação descreve os endpoints disponíveis para integração com o sistema web da Drogaria, desenvolvido em Laravel. A API pode ser utilizada por qualquer linguagem de programação capaz de realizar requisições HTTP.",
        "contact": {
            "name": "Bruno",
            "email": "bruno.facomp2016@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/orcamentos": {
            "get": {
                "tags": [
                    "Orçamentos"
                ],
                "summary": "Lista todos os orçamentos com seus itens, cliente e vendedor",
                "operationId": "ae29be4b0df790db04435c651e221590",
                "responses": {
                    "200": {
                        "description": "Lista de orçamentos retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "cliente_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "vendedor_id": {
                                                "type": "integer",
                                                "example": 2
                                            },
                                            "data_orcamento": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2024-05-01"
                                            },
                                            "cliente": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "nome": {
                                                        "type": "string",
                                                        "example": "Cliente A"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "vendedor": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "nome": {
                                                        "type": "string",
                                                        "example": "Vendedor B"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "itens": {
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 10
                                                        },
                                                        "produto_id": {
                                                            "type": "integer",
                                                            "example": 5
                                                        },
                                                        "quantidade": {
                                                            "type": "integer",
                                                            "example": 3
                                                        },
                                                        "preco_unitario": {
                                                            "type": "number",
                                                            "format": "float",
                                                            "example": 19.99
                                                        },
                                                        "desconto": {
                                                            "type": "number",
                                                            "format": "float",
                                                            "example": 2.5
                                                        },
                                                        "produto": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "integer",
                                                                    "example": 5
                                                                },
                                                                "nome": {
                                                                    "type": "string",
                                                                    "example": "Produto X"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Orçamentos"
                ],
                "summary": "Cria um novo orçamento",
                "operationId": "1a49dc5722eddca129869a3e70088703",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "cliente_id",
                                    "vendedor_id",
                                    "data_orcamento",
                                    "itens"
                                ],
                                "properties": {
                                    "cliente_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "vendedor_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "data_orcamento": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-05-01"
                                    },
                                    "itens": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "produto_id",
                                                "quantidade"
                                            ],
                                            "properties": {
                                                "produto_id": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "quantidade": {
                                                    "type": "integer",
                                                    "example": 3
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Orçamento criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "cliente_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "vendedor_id": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "data_orcamento": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2024-05-01"
                                        },
                                        "cliente": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "nome": {
                                                    "type": "string",
                                                    "example": "Cliente A"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "vendedor": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "nome": {
                                                    "type": "string",
                                                    "example": "Vendedor B"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "itens": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 10
                                                    },
                                                    "produto_id": {
                                                        "type": "integer",
                                                        "example": 5
                                                    },
                                                    "quantidade": {
                                                        "type": "integer",
                                                        "example": 3
                                                    },
                                                    "preco_unitario": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 19.99
                                                    },
                                                    "desconto": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 2.5
                                                    },
                                                    "produto": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 5
                                                            },
                                                            "nome": {
                                                                "type": "string",
                                                                "example": "Produto X"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação.",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "errors": {
                                        "campo": [
                                            "mensagem"
                                        ]
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}